fix: Fixes for some bugs with Instagram Stories on Android (#856)
* [android] more efficient base64 processing

previously, on large base64 files, there would be a stackOverflow calling
`getSchemeSpecificPart`. Given that base64 data is very structured, we
can get it with offsets directly. This improves the performance to a
point where it's usable with large base64 data.

fixes #852

* [android] improve speed to open intent

remove an extra call to `super.open` that was doing the exact same
things with the exact same data for the exact same result. Once we've
called `super.open` once, we can exit from this function.

fixes #853

* [android] don't overwrite flags before sending intent

when using `shareSingle`, the flags would be overwritten before sending
the intent. This would result in `FLAG_GRANT_READ_URI_PERMISSION`, for
instance, not being set on the intent any more, which means any linked
files would not be accessible by the target application.

fixes #854

* [android] ensure instagram stick is alway granted read permission

There were certain code paths previously where the instagram sticker was
not flagged with FLAG_GRANT_READ_URI_PERMISSION, which meant that
Instagram was unable to open the sticker. This moves the logic around
granting permission next to the logic about adding the sticker to the
intent. It is duplicated, but it was going to be duplicated _anyway_. At
least if it's here, we're only looking into the options deeply one time.

fixes #855